backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
python-multipart==0.0.6
arabic-reshaper==3.0.0
python-bidi==0.4.2
psycopg2-binary==2.9.9